requirements.txt
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.2
Pillow==8.1.1
pyparsing==2.4.7
influxdb==5.3.0
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.2
six==1.15.0
wget==3.2
joblib==0.17.0
cython
SciencePlots
seaborn
sklearn-extensions
tqdm